split utils into separate files: errorhandler and httprequests
